Bug fix, red description
After the Japan GP an error came up. The current F1 championship leader can win the championship in Qatar (the next race) this created an error when collecting the standings from a few races before. This has now been fixed. The result will still give the result of the current championship leader LOSING the championship.